Issue 9: d1 templates, styles, scripts, calendar view tabs and heads, edit view multi reference, recurrence, reminder

+++ modules/ITS4YouCalendar/actions/MassDelete.php
@@ -28,7 +28,7 @@ class ITS4YouCalendar_MassDelete_Action extends Vtiger_MassDelete_Action
         $cvId = $request->get('viewname');
 
         foreach ($recordIds as $recordId) {
-            if (Users_Privileges_Model::isPermitted($moduleName, 'Delete', $recordId)) {
+            if (!empty($recordId) && isRecordExists($recordId) && Users_Privileges_Model::isPermitted($moduleName, 'Delete', $recordId)) {
                 $recordModel = Vtiger_Record_Model::getInstanceById($recordId, $moduleModel);
                 $recordModel->delete();
